+++ iframe2.html
@@ -13,7 +13,7 @@
             <h1>Page d'affichage dynamique de l'iframe de l'application</h1>
             <p>Ci-dessous s'affiche l'iframe de l'application</p>
         </div>
-        <script src="http://localhost:8000/static/iframe.min.js" id="lvao_script" data-params='{"first_dir":"jai","direction":"jai"}'></script>
+        <script src="http://localhost:8000/static/iframe.js" id="lvao_script" data-params='{"first_dir":"jai","direction":"jai"}'></script>
         <div>
             <p>Fin de l'affichage de l'iframe de l'application</p>
         </div>
