+++ package.json
@@ -11,6 +11,7 @@
         "static/to_compile/entrypoints/qfdmo.css",
         "static/to_compile/entrypoints/qfdmo.ts",
         "static/to_compile/entrypoints/iframe_resizer.ts",
+        "static/to_compile/entrypoints/iframe.ts",
         "static/to_compile/entrypoints/admin-map-widget.ts"
     ],
     "targets": {
